requirements.txt
amqp==2.6.1
asgiref==3.3.4
auditlog3==1.0.1
billiard==3.6.4.0
celery==4.4.7
certifi==2020.12.5
Django==3.2
django-cors-headers==3.7.0
django-dotenv==1.4.2
django-filter==2.4.0
django-redis==5.0.0
django-rest-framework-nested==0.0.1
django-user-agents==0.4.0
django-utils-six==2.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
kombu==4.6.11
psycopg2==2.8.6
psycopg2-binary==2.8.6
PyJWT==1.7.1
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
sentry-sdk==1.0.0
six==1.16.0
sqlparse==0.4.1
ua-parser==0.10.0
urllib3==1.26.4
user-agents==2.2.0
vine==1.3.0
